@@ -7,35 +7,37 @@
     "doc": "docs"
   },
   "devDependencies": {
-    "babel-cli": "^6.23.0",
-    "babel-core": "^6.23.1",
-    "babel-loader": "^6.3.2",
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^6.4.1",
     "babel-plugin-react-transform": "^2.0.2",
     "babel-plugin-transform-flow-strip-types": "^6.22.0",
     "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-preset-es2015": "^6.22.0",
-    "babel-preset-react": "^6.23.0",
-    "babel-preset-stage-2": "^6.22.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
     "babel-runtime": "^6.23.0",
     "babel-tape-runner": "^2.0.1",
-    "babel-types": "^6.23.0",
+    "babel-types": "^6.24.1",
     "babelify": "^7.3.0",
     "blue-tape": "^1.0.0",
-    "eslint": "^3.16.1",
+    "create-react-class": "^15.5.2",
+    "eslint": "^3.19.0",
     "eslint-config-shakacode": "^14.1.1",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^6.10.0",
-    "flow-bin": "^0.40.0",
-    "jsdom": "^9.11.0",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "eslint-plugin-react": "^6.10.3",
+    "flow-bin": "^0.43.1",
+    "jsdom": "^9.12.0",
+    "prop-types": "^15.5.7",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
     "react-transform-hmr": "^1.0.4",
     "redux": "^3.6.0",
-    "release-it": "^2.5.4",
+    "release-it": "^2.7.1",
     "tap-spec": "^4.1.1",
     "tape": "^4.6.3",
-    "webpack": "^2.3.1"
+    "webpack": "^2.3.3"
   },
   "peerDependencies": {
     "babel-runtime": ">= 6",
