@@ -8,56 +8,57 @@
   },
   "main": "js/app.js",
   "dependencies": {
-    "autoprefixer": "^6.7.6",
-    "babel-cli": "^6.23.0",
-    "babel-core": "^6.23.1",
-    "babel-loader": "^6.3.2",
+    "autoprefixer": "^6.7.7",
+    "babel-cli": "^6.24.1",
+    "babel-core": "^6.24.1",
+    "babel-loader": "^6.4.1",
     "babel-polyfill": "^6.23.0",
-    "babel-preset-es2015": "^6.22.0",
-    "babel-preset-react": "^6.23.0",
-    "babel-preset-stage-2": "^6.22.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-2": "^6.24.1",
     "babel-runtime": "^6.23.0",
-    "css-loader": "^0.26.2",
+    "css-loader": "^0.28.0",
     "es5-shim": "^4.5.9",
     "expose-loader": "^0.7.3",
-    "extract-text-webpack-plugin": "^2.0.0",
-    "file-loader": "^0.10.1",
-    "history": "^4.5.1",
+    "extract-text-webpack-plugin": "^2.1.0",
+    "file-loader": "^0.11.1",
+    "history": "^4.6.1",
     "imports-loader": "^0.7.1",
-    "jquery": "^3.1.1",
+    "jquery": "^3.2.1",
     "jquery-ujs": "^1.2.2",
-    "loader-utils": "^1.0.2",
+    "loader-utils": "^1.1.0",
     "lodash": "^4.17.4",
     "node-libs-browser": "^2.0.0",
-    "node-sass": "^4.5.0",
-    "postcss-loader": "^1.3.2",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
+    "node-sass": "^4.5.2",
+    "postcss-loader": "^1.3.3",
+    "react": "^15.5.4",
+    "react-dom": "^15.5.4",
     "react-on-rails": "file:../../..",
-    "react-redux": "^5.0.3",
-    "react-router": "^3.0.2",
+    "react-proptypes": "^0.0.1",
+    "react-redux": "^5.0.4",
+    "react-router": "3.0.5",
     "redux": "^3.6.0",
     "redux-thunk": "^2.2.0",
-    "resolve-url-loader": "^2.0.0",
-    "sass-loader": "^6.0.2",
-    "sass-resources-loader": "1.2.0",
-    "style-loader": "^0.13.2",
+    "resolve-url-loader": "^2.0.2",
+    "sass-loader": "^6.0.3",
+    "sass-resources-loader": "1.2.1",
+    "style-loader": "^0.16.1",
     "url-loader": "^0.5.8",
-    "webpack": "^2.3.1"
+    "webpack": "^2.3.3"
   },
   "devDependencies": {
     "babel-plugin-react-transform": "^2.0.2",
-    "body-parser": "^1.16.1",
-    "eslint": "^3.16.1",
+    "body-parser": "^1.17.1",
+    "eslint": "^3.19.0",
     "eslint-config-shakacode": "^14.1.1",
     "eslint-plugin-import": "^2.2.0",
     "eslint-plugin-jsx-a11y": "^4.0.0",
-    "eslint-plugin-react": "^6.10.0",
-    "fbjs": "^0.8.9",
-    "jsdom": "^9.11.0",
+    "eslint-plugin-react": "^6.10.3",
+    "fbjs": "^0.8.12",
+    "jsdom": "^9.12.0",
     "react-transform-hmr": "^1.0.4",
     "tape": "^4.6.3",
-    "webpack-dev-server": "^2.4.1"
+    "webpack-dev-server": "^2.4.2"
   },
   "browser": {
     "fs": false
